Code review

+++ bot.py
@@ -32,13 +32,16 @@ async def on_shutdown(dispatcher):
 
 
 smileys = [
-    '😀', '😃', '😄', '😁', '😆', '😅',
-    '🙂', '😂', '🤣', '😊', '😇', '🙃',
-    '😉', '😌', '😘', '😍', '😗', '😙',
-    '😚', '😋', '😛', '😜', '😝', '🤪',
-    '🤨', '🤢', '🤮', '🤧', '🥱', '🥶',
-    '😤', '😱', '😓', '😡', '🥵', '🤬',
-]
+    "😊", "😀", "🤪", "😍", "😅",
+    "😆", "😉", "😌", "😎", "😏",
+    "🤔", "😒", "😔", "😕", "😖",
+    "🤢", "😟", "😠", "😡", "😢",
+    "😣", "😥", "😪", "😫", "😴"]
+
+"""списки для кнопок"""
+buttons_menu = ["Статистика", "Выбрать смайлик"]
+admin_menu = ["Кол-во новых пользователей за неделю", "Общее кол-во пользовтелей", "Статистика за день",
+              "Статистика за неделю", "Статистика за месяц", "Выйти"]
 
 
 @dp.message_handler(commands=['start'])
@@ -47,20 +50,34 @@ async def start(message: types.Message):
     user_exists = await database.checkUser(str(message.from_user.id))
     if not user_exists:
         await database.createUser(message.from_user.id, message.from_user.username)
+    await message.answer('Выбери что тебя интересует', reply_markup=show_button(buttons_menu))
 
 
-@dp.message_handler(commands=["stat"])
+@dp.message_handler(text=["Статистика"])
 async def statisticUser(message: types.Message):
     user_id = message.from_user.id  # ID чата
-    pathToPicture = statistics.analiticData(user_id)  # путь к картинке со статой
-    await message.answer("Ваша статистика за неделю")
-    photo = InputFile(pathToPicture)
-    await bot.send_photo(chat_id=message.chat.id, photo=photo)
-    os.remove(pathToPicture)  # удаляем файл с картинкой
+    pathToPicture = await statistics.analiticData(user_id)  # путь к картинке со статой
+    if pathToPicture != "absent":
+        await message.answer("Ваша статистика за всё время")
+        photo = InputFile(pathToPicture)
+        await bot.send_photo(chat_id=message.chat.id, photo=photo)
+        os.remove(pathToPicture)  # удаляем файл с картинкой
+    else:
+        await message.answer("Статистика отсутствует. Вы еще ни разу не вводили смайлики.")
 
 
-def show_button(list_emoji):
-    keyboard = InlineKeyboardMarkup(row_width=6)
+def show_button(list_menu):
+    """Принимает список и превращает его в кнопки"""
+    """создает кнопки для меню"""
+    keyboard = types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard.add(*list_menu)
+    return keyboard
+
+
+def show_inline_button(list_emoji):
+    """Принимает список и превращает его в инлайн кнопки кнопки"""
+    "создает инлайн кнопки для показа смайликов"
+    keyboard = InlineKeyboardMarkup(row_width=5)
     buttons = [InlineKeyboardButton(smiley, callback_data=smiley) for smiley in list_emoji]
     keyboard.add(*buttons)
     return keyboard
@@ -70,9 +87,9 @@ def add_checkmark(lst, variable):
     return [elem + "✅" if elem == variable else elem for elem in lst]
 
 
-@dp.message_handler(commands=['emoji'])
+@dp.message_handler(text=["Выбрать смайлик"])
 async def show_emoji(message: types.Message):
-    await message.reply('Выберите смайлик:', reply_markup=show_button(smileys))
+    await message.reply('Выберите смайлик:', reply_markup=show_inline_button(smileys))
 
 
 @dp.callback_query_handler()
@@ -81,10 +98,56 @@ async def button(callback_query: types.CallbackQuery):
     await query.answer()
     new_emoji_list = add_checkmark(smileys, query.data)
     await bot.answer_callback_query(callback_query.id)
-    await query.message.edit_text('Выбранный смайлик ✅', reply_markup=show_button(new_emoji_list))
-    date_day = date.today()
-    date_name = calendar.day_name[date_day.weekday()]
-    await database.addOrChangeSmile(callback_query.from_user.id, date_name, '' if '✅' in query.data else query.data)
+    await query.message.edit_text('Выбранный смайлик ✅', reply_markup=show_inline_button(new_emoji_list))
+    date_day = str(date.today())
+    await database.addOrChangeSmile(callback_query.from_user.id, date_day, query.data)
+
+
+@dp.message_handler(commands=['admin'])
+async def admin(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        await message.answer('Выполнен вход в админ-панель', reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Кол-во новых пользователей за неделю"])
+async def stat_new_week(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        await message.answer(f'Кол-во новых пользователей за неделю: {await database.getCountNewUsers()}',
+                             reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Общее кол-во пользовтелей"])
+async def stat_all(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        await message.answer(f'Общее кол-во пользовтелей: {await database.getCountAllUsers()}',
+                             reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Статистика за день"])
+async def stat_day(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        info = await database.getStatAdmin(1)
+        await message.answer(f'Статистика за день: {info}', reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Статистика за неделю"])
+async def stat_week(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        info = await database.getStatAdmin(7)
+        await message.answer(f'Статистика за неделю: \n{" ".join(info)}', reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Статистика за месяц"])
+async def stat_month(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        info = await database.getStatAdmin(30)
+        await message.answer(f'Статистика за месяц: {info}', reply_markup=show_button(admin_menu))
+
+
+@dp.message_handler(text=["Выйти"])
+async def admin_exit(message: types.Message):
+    if message.from_user.id == config.ADMIN_ID:
+        await message.reply('Выход из админ-панели', reply_markup=show_button(buttons_menu))
 
 
 if __name__ == '__main__':
